Grading system + request form

+++ LibeRate/LibeRate/Models/Book.cs
@@ -10,10 +10,10 @@ namespace LibeRate.Models
         public string Title { get; set; }
         public string Author { get; set; }
         public string ImageURL { get; set; }
-        public int DifficultyRating { get; set; }
+        public float DifficultyRating { get; set; }
 
         public Book() { }
-        public Book(string id, string title, string author, string imageURL, int difficultyRating)
+        public Book(string id, string title, string author, string imageURL, float difficultyRating)
         {
             Id = id;
             Title = title;
